+++ packages/navigator/src/browser/navigator-contribution.ts
@@ -38,7 +38,6 @@ import { FileDownloadCommands } from '@theia/filesystem/lib/browser/download/fil
 import {
     Command,
     CommandRegistry,
-    DisposableCollection,
     isOSX,
     MenuModelRegistry,
     MenuPath,
@@ -257,17 +256,6 @@ export class FileNavigatorContribution extends AbstractViewContribution<FileNavi
                 this.openView({ activate: true });
             }
         }
-    }
-
-    async onStart(app: FrontendApplication): Promise<void> {
-        this.workspacePreferences.ready.then(() => {
-            this.updateAddRemoveFolderActions(this.menuRegistry);
-            this.workspacePreferences.onPreferenceChanged(change => {
-                if (change.preferenceName === 'workspace.supportMultiRootWorkspace') {
-                    this.updateAddRemoveFolderActions(this.menuRegistry);
-                }
-            });
-        });
     }
 
     async initializeLayout(app: FrontendApplication): Promise<void> {
@@ -549,6 +537,14 @@ export class FileNavigatorContribution extends AbstractViewContribution<FileNavi
             label: nls.localizeByDefault('Close All'),
             order: 'c'
         });
+
+        registry.registerMenuAction(NavigatorContextMenu.WORKSPACE, {
+            commandId: FileNavigatorCommands.ADD_ROOT_FOLDER.id,
+            label: WorkspaceCommands.ADD_FOLDER.label
+        });
+        registry.registerMenuAction(NavigatorContextMenu.WORKSPACE, {
+            commandId: WorkspaceCommands.REMOVE_FOLDER.id
+        });
     }
 
     override registerKeybindings(registry: KeybindingRegistry): void {
@@ -710,20 +706,6 @@ export class FileNavigatorContribution extends AbstractViewContribution<FileNavi
     async refreshWorkspace(): Promise<void> {
         const { model } = await this.widget;
         await model.refresh();
-    }
-
-    private readonly toDisposeAddRemoveFolderActions = new DisposableCollection();
-    private updateAddRemoveFolderActions(registry: MenuModelRegistry): void {
-        this.toDisposeAddRemoveFolderActions.dispose();
-        if (this.workspacePreferences['workspace.supportMultiRootWorkspace']) {
-            this.toDisposeAddRemoveFolderActions.push(registry.registerMenuAction(NavigatorContextMenu.WORKSPACE, {
-                commandId: FileNavigatorCommands.ADD_ROOT_FOLDER.id,
-                label: WorkspaceCommands.ADD_FOLDER.label!
-            }));
-            this.toDisposeAddRemoveFolderActions.push(registry.registerMenuAction(NavigatorContextMenu.WORKSPACE, {
-                commandId: WorkspaceCommands.REMOVE_FOLDER.id
-            }));
-        }
     }
 
 }